+++ semantic-conventions-genai/src/index-incubating.ts
@@ -1,0 +1,18 @@
+/*
+ * Copyright The OpenTelemetry Authors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+/* eslint-disable no-restricted-syntax --
+ * These re-exports are only of constants, only two-levels deep, and
+ * should not cause problems for tree-shakers.
+ */
+
+// Incubating export also contains stable constants in order to maintain
+// backward compatibility between minor version releases
+export * from './stable_attributes';
+export * from './stable_metrics';
+export * from './stable_events';
+export * from './experimental_attributes';
+export * from './experimental_metrics';
+export * from './experimental_events';